Usual mistake of not casting an operand to float.

+++ include/Bins.hpp
@@ -38,10 +38,10 @@ template< typename T > std::vector< unsigned int > * getNrSamplesPerBin(const As
     std::fill(itemsPerBin.begin(), itemsPerBin.end(), 0);
     std::fill(offsetPerBin.begin(), offsetPerBin.end(), 0);
     for ( unsigned int i = 0; i < period; i++ ) {
-      float samplePhase = (i / period);
+      float samplePhase = (static_cast< float >(i) / period);
 
       for ( unsigned int bin = 0; bin < obs.getNrBins(); bin++ ) {
-        if ( samplePhase - ((bin + 0.5) / obs.getNrBins()) < 1.0f / obs.getNrBins() ) {
+        if ( samplePhase - ((bin + 0.5f) / obs.getNrBins()) < 1.0f / obs.getNrBins() ) {
           itemsPerBin[bin] += 1;
           break;
         }
